코드 리뷰 반영 : Channel 클래스 생성, View 클래스 재정의
