+++ src/main/java/racingcar/external/OuputChannel.java
@@ -1,0 +1,8 @@
+package racingcar.external;
+
+public class OuputChannel implements Output {
+    @Override
+    public void out(String s) {
+        System.out.println(s);
+    }
+}
