+++ src/main/java/racingcar/external/ResultView.java
@@ -1,25 +1,24 @@
 package racingcar.external;
 
-import racingcar.model.Car;
+import racingcar.model.Cars;
 
 import java.util.stream.Stream;
 
 import static java.util.stream.Collectors.joining;
 
-public class ResultView implements Output {
+public class ResultView {
 
     private static final String MOVE_DASH = "-";
+    private static final String RACE_TREM = "********************";
 
-    @Override
-    public void out(String s) {
-        System.out.println(s);
+    public static void output(OuputChannel outputChnnel, Cars cars) {
+        for (int position : cars.getPositions()) {
+            outputChnnel.out(getMoveDash(position));
+        }
+        outputChnnel.out(RACE_TREM);
     }
 
-    public void out(Car car) {
-        out(getMoveDash(car.getPosition()));
-    }
-
-    public String getMoveDash(int position) {
+    public static String getMoveDash(int position) {
         return Stream.generate(() -> MOVE_DASH)
                 .limit(position)
                 .collect(joining());
